ebslib.date - correctly mix events into the estimate
The Estimate command included all future event hours in the estimate;
the time it takes to complete tasks might not encounter some (or any)
future events, so this is not correct.

Adopt a recursive approach, using tasks only for the estimates, then
calculating a new ship date using the hours of events falling within
the timespan.  This continues recursively until the dates converge.

Also add the ``'--exponent'`` argument to the Estimate command,
allowing the user to ask for 10 ** N rounds of simulation for N
>= 2 (2 is both the default, and the minimum possible number of
rounds).