ebslib.task.Task - refactor 'completed' to be an attribute
Currently, any task with an actual of zero is assume to have not
been completed.  As a first step in addressing this limitiation,
refactor the ``completed`` @property to be an attribute.  To
maintain backwards compatibility, the __init__ default is None,
and when None, the stored value will be derived from the actual.