test_date: fix tests on non-work days
Some tests assume that every day is a work day.  Running the tests
on a work day would succeed, but on a non-work day, a couple of
tests are failing.

Add a little function to get the next work date and update the
broken tests to use this function instead of assuming that the
current date is a work day.